requirements.txt
fleep==1.0.1
Pillow==10.1.0
